Changed the properties names and added necessary properties for a MySQL and H2 database backends.
Added another data source, tx manager, and entity manager factory
for MySQL.  Added the entity-manager-factory-ref and
transaction-manager-ref to jpa:repositories entry in the spring
application context.

Added a MySQL persistence unit definition to the persistence.xml.

Added the unitName to the PersonDao and the
PersonRepositoryOverrideImpl.

At this point we are not using the second MySQL persistence unit
but we are ready and configured for it.